# [chore] Refactor allocation strategies