--- a/cmd/otel-allocator/allocation/consistent_hashing_test.go
+++ b/cmd/otel-allocator/allocation/consistent_hashing_test.go
@@ -22,7 +22,7 @@ import (
 
 func TestCanSetSingleTarget(t *testing.T) {
 	cols := MakeNCollectors(3, 0)
-	c := newConsistentHashingAllocator(logger)
+	c := newAllocator(logger, newConsistentHashingStrategy())
 	c.SetCollectors(cols)
 	c.SetTargets(MakeNNewTargets(1, 3, 0))
 	actualTargetItems := c.TargetItems()
@@ -38,7 +38,7 @@ func TestRelativelyEvenDistribution(t *testing.T) {
 	cols := MakeNCollectors(numCols, 0)
 	var expectedPerCollector = float64(numItems / numCols)
 	expectedDelta := (expectedPerCollector * 1.5) - expectedPerCollector
-	c := newConsistentHashingAllocator(logger)
+	c := newAllocator(logger, newConsistentHashingStrategy())
 	c.SetCollectors(cols)
 	c.SetTargets(MakeNNewTargets(numItems, 0, 0))
 	actualTargetItems := c.TargetItems()
@@ -52,7 +52,7 @@ func TestRelativelyEvenDistribution(t *testing.T) {
 
 func TestFullReallocation(t *testing.T) {
 	cols := MakeNCollectors(10, 0)
-	c := newConsistentHashingAllocator(logger)
+	c := newAllocator(logger, newConsistentHashingStrategy())
 	c.SetCollectors(cols)
 	c.SetTargets(MakeNNewTargets(10000, 10, 0))
 	actualTargetItems := c.TargetItems()
@@ -77,7 +77,7 @@ func TestNumRemapped(t *testing.T) {
 	numFinalCols := 16
 	expectedDelta := float64((numFinalCols - numInitialCols) * (numItems / numFinalCols))
 	cols := MakeNCollectors(numInitialCols, 0)
-	c := newConsistentHashingAllocator(logger)
+	c := newAllocator(logger, newConsistentHashingStrategy())
 	c.SetCollectors(cols)
 	c.SetTargets(MakeNNewTargets(numItems, numInitialCols, 0))
 	actualTargetItems := c.TargetItems()
@@ -106,7 +106,7 @@ func TestNumRemapped(t *testing.T) {
 
 func TestTargetsWithNoCollectorsConsistentHashing(t *testing.T) {
 
-	c := newConsistentHashingAllocator(logger)
+	c := newAllocator(logger, newConsistentHashingStrategy())
 
 	// Adding 10 new targets
 	numItems := 10