Allocation strategies are currently messy and duplicate a lot of logic. This is an attempt at making them more DRY.

* Factored out the bookkeeping into a new Allocator struct. It keeps track of targets, collectors, allocation, metrics, etc.
* Added a new Strategy interface that strategies need to implement. Right now, the minimum needed is just a method that takes a collector map and a target, and returns the assigned collector.
* Added some generic allocator tests that run for all strategies.

I'm not entirely happy with the end result, and I've left some TODOs in the code I'd like to address in a follow-up. I haven't included those changes in this PR in the interest of keeping it simpler.