@swiatekm-sumo 2024-05-06
> yeah this is super close to what i tried to do a year or two ago, and am definitely in favor of this change. The trouble I ran into back then was how you give the right context to the strategy without duplicating memory. i.e. it's not great if the strategy has to keep its own collector map and the allocator struct does too. Overall, I think the approach makes sense. I'll give it a better read later today!

Is that a realistic problem? The collector count is going to be less than 100 in almost all practical circumstances, and we're only storing pointers anyway, so the memory cost is small. None of my strategy implementations *need* to store collectors anyway, the `per-node` strategy only does it for performance reasons.

In a follow-up, I'd like to restrict what data actually gets passed to the strategy in the `SetCollectors` call. Right now, the strategy can assume some of this data is immutable until the next call (like the name or node), but not all of it (`NumTargets` can change). I'd like to express this more directly in the API instead of relying on convention.