@swiatekm-sumo 2024-05-06
> I was worried about the labelsets cardinality, but yeah if we're just using some pointers for only the collectors we should be alright. Either way, I would love to see some benchmarks to ensure that's the case. It feels like we've had some brush ups with memory and the TA recently.

Ran the `Benchmark_Setting` benchmark, which looks the most appropriate for my changes. Tried others and saw no change.

```
                                                             │    old.txt    │               new.txt                │
                                                             │    sec/op     │    sec/op     vs base                │
_Setting/consistent-hashing_num_cols_100_num_jobs_100-32        6.042µ ±  3%   6.054µ ±  5%        ~ (p=0.481 n=10)
_Setting/consistent-hashing_num_cols_100_num_jobs_1000-32       51.12µ ±  4%   50.71µ ±  6%        ~ (p=0.529 n=10)
_Setting/consistent-hashing_num_cols_100_num_jobs_10000-32      723.1µ ±  5%   726.4µ ±  1%        ~ (p=0.579 n=10)
_Setting/consistent-hashing_num_cols_100_num_jobs_100000-32     8.688m ± 12%   8.763m ±  3%        ~ (p=0.529 n=10)
_Setting/consistent-hashing_num_cols_1000_num_jobs_100-32       48.88µ ±  7%   49.89µ ±  6%        ~ (p=0.218 n=10)
_Setting/consistent-hashing_num_cols_1000_num_jobs_1000-32      103.6µ ±  3%   105.2µ ±  2%        ~ (p=0.684 n=10)
_Setting/consistent-hashing_num_cols_1000_num_jobs_10000-32     781.2µ ±  5%   781.5µ ±  2%        ~ (p=0.853 n=10)
_Setting/consistent-hashing_num_cols_1000_num_jobs_100000-32    8.701m ±  8%   9.831m ± 12%  +12.99% (p=0.007 n=10)
_Setting/per-node_num_cols_100_num_jobs_100-32                 42.934µ ±  1%   5.933µ ±  5%  -86.18% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_1000-32                175.94µ ±  1%   50.84µ ±  6%  -71.10% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_10000-32               1723.0µ ±  3%   726.5µ ±  6%  -57.84% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_100000-32              28.073m ± 20%   8.726m ±  4%  -68.92% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100-32                346.49µ ±  1%   48.89µ ±  7%  -85.89% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_1000-32                496.6µ ±  1%   105.4µ ±  3%  -78.77% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_10000-32              2065.8µ ±  2%   768.4µ ±  3%  -62.80% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100000-32             27.328m ± 15%   8.874m ± 10%  -67.53% (p=0.000 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_100-32            5.885µ ±  4%   5.971µ ±  4%        ~ (p=0.089 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_1000-32           51.27µ ±  6%   50.84µ ±  8%        ~ (p=0.579 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_10000-32          701.0µ ±  5%   703.9µ ±  3%        ~ (p=0.190 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_100000-32         8.838m ± 10%   8.635m ± 16%        ~ (p=0.063 n=10)
_Setting/least-weighted_num_cols_1000_num_jobs_100-32           49.83µ ±  4%   49.19µ ±  3%        ~ (p=0.469 n=10)
_Setting/least-weighted_num_cols_1000_num_jobs_1000-32          105.2µ ±  2%   105.1µ ±  2%        ~ (p=0.579 n=10)
_Setting/least-weighted_num_cols_1000_num_jobs_10000-32         761.8µ ±  5%   783.8µ ±  5%        ~ (p=0.739 n=10)
_Setting/least-weighted_num_cols_1000_num_jobs_100000-32        8.951m ±  8%   8.805m ±  7%        ~ (p=0.315 n=10)
geomean                                                         473.1µ         302.6µ        -36.04%

                                                             │    old.txt     │                new.txt                │
                                                             │      B/op      │    B/op     vs base                   │
_Setting/consistent-hashing_num_cols_100_num_jobs_100-32           192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_1000-32          192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_10000-32         192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_100000-32        192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_100-32          192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_1000-32         192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_10000-32        192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_100000-32       192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/per-node_num_cols_100_num_jobs_100-32                   27842.5 ± 0%   192.0 ± 0%   -99.31% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_1000-32                  85440.0 ± 0%   192.0 ± 0%   -99.78% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_10000-32                661422.0 ± 0%   192.0 ± 0%   -99.97% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_100000-32              6421405.5 ± 0%   192.0 ± 0%  -100.00% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100-32                 299523.5 ± 0%   192.0 ± 0%   -99.94% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_1000-32                357104.0 ± 0%   192.0 ± 0%   -99.95% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_10000-32               933133.0 ± 0%   192.0 ± 0%   -99.98% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100000-32             6693000.0 ± 0%   192.0 ± 0%  -100.00% (p=0.000 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_100-32               192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_1000-32              192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_10000-32             192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_100000-32            192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_100-32              192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_1000-32             192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_10000-32            192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_100000-32           192.0 ± 0%   192.0 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                                          2.639Ki        192.0        -92.90%
¹ all samples are equal

                                                             │     old.txt     │                new.txt                │
                                                             │    allocs/op    │ allocs/op   vs base                   │
_Setting/consistent-hashing_num_cols_100_num_jobs_100-32            4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_1000-32           4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_10000-32          4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_100_num_jobs_100000-32         4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_100-32           4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_1000-32          4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_10000-32         4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/consistent-hashing_num_cols_1000_num_jobs_100000-32        4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/per-node_num_cols_100_num_jobs_100-32                    222.000 ± 0%   4.000 ± 0%   -98.20% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_1000-32                  1122.000 ± 0%   4.000 ± 0%   -99.64% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_10000-32                10122.000 ± 0%   4.000 ± 0%   -99.96% (p=0.000 n=10)
_Setting/per-node_num_cols_100_num_jobs_100000-32              100122.000 ± 0%   4.000 ± 0%  -100.00% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100-32                  1185.000 ± 0%   4.000 ± 0%   -99.66% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_1000-32                 2085.000 ± 0%   4.000 ± 0%   -99.81% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_10000-32               11085.000 ± 0%   4.000 ± 0%   -99.96% (p=0.000 n=10)
_Setting/per-node_num_cols_1000_num_jobs_100000-32             101085.000 ± 0%   4.000 ± 0%  -100.00% (p=0.000 n=10)
_Setting/least-weighted_num_cols_100_num_jobs_100-32                4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_1000-32               4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_10000-32              4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_100_num_jobs_100000-32             4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_100-32               4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_1000-32              4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_10000-32             4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
_Setting/least-weighted_num_cols_1000_num_jobs_100000-32            4.000 ± 0%   4.000 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                                             44.15        4.000        -90.94%
```

So `consistent-hashing` is slightly slower and `per-node` is much faster and uses much less memory.